[FIX] calendar: hide res model name while creating an event from another model
Before this commit when you create an event from a lead, a button appears to jump back to the lead but while you're creating the record, it already appears but empty.
This commit hides the button in case the res_model_name is null, which happens while we are creating the event and before hitting save.

closes odoo/odoo#156884

Task: 3747059
X-original-commit: 8fd5c00ebd3ae4f978bbc6ef6efbf33488d896ce
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>